TrainsLogger -> ClearMLLogger
