+++ docker/msdp/Dockerfile.msdp-apex-vision
@@ -16,4 +16,4 @@ RUN pip install --upgrade --no-cache-dir albumentations \
                                          opencv-python \
                                          py_config_runner \
                                          pillow \
-                                         trains
+                                         clearml
