+++ examples/contrib/cifar10/README.md
@@ -1,6 +1,7 @@
 # CIFAR10 Example with Ignite
 
-In this example, we show how to use *Ignite* to train a neural network:
+In this example, we show how to use _Ignite_ to train a neural network:
+
 - on 1 or more GPUs or TPUs
 - compute training/validation metrics
 - log learning rate, metrics etc
@@ -8,10 +9,10 @@ In this example, we show how to use *Ignite* to train a neural network:
 
 Configurations:
 
-* [x] single GPU
-* [x] multi GPUs on a single node
-* [x] multi GPUs on multiple nodes
-* [x] TPUs on Colab
+- [x] single GPU
+- [x] multi GPUs on a single node
+- [x] multi GPUs on multiple nodes
+- [x] TPUs on Colab
 
 ## Requirements:
 
@@ -20,21 +21,24 @@ Configurations:
 - [tqdm](https://github.com/tqdm/tqdm/): `pip install tqdm`
 - [tensorboardx](https://github.com/lanpa/tensorboard-pytorch): `pip install tensorboardX`
 - [python-fire](https://github.com/google/python-fire): `pip install fire`
-- Optional: [trains](https://github.com/allegroai/trains): `pip install trains`
+- Optional: [clearml](https://github.com/allegroai/clearml): `pip install clearml`
 
 ## Usage:
 
 Run the example on a single GPU:
+
 ```bash
 python main.py run
 ```
 
 For details on accepted arguments:
+
 ```bash
 python main.py run -- --help
 ```
 
 If user would like to provide already downloaded dataset, the path can be setup in parameters as
+
 ```bash
 --data_path="/path/to/cifar10/"
 ```
@@ -44,11 +48,14 @@ If user would like to provide already downloaded dataset, the path can be setup 
 #### Single node, multiple GPUs
 
 Let's start training on a single node with 2 gpus:
+
 ```bash
 # using torch.distributed.launch
 python -u -m torch.distributed.launch --nproc_per_node=2 --use_env main.py run --backend="nccl"
 ```
-or 
+
+or
+
 ```bash
 # using function spawn inside the code
 python -u main.py run --backend="nccl" --nproc_per_node=2
@@ -59,28 +66,29 @@ python -u main.py run --backend="nccl" --nproc_per_node=2
 Please, make sure to have Horovod installed before running.
 
 Let's start training on a single node with 2 gpus:
+
 ```bash
 # horovodrun
 horovodrun -np=2 python -u main.py run --backend="horovod"
 ```
-or 
+
+or
+
 ```bash
 # using function spawn inside the code
 python -u main.py run --backend="horovod" --nproc_per_node=2
 ```
 
-
 #### Colab, on 8 TPUs
 
-
 Same code can be run on TPUs: [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1E9zJrptnLJ_PKhmaP5Vhb6DTVRvyrKHx)
-
 
 #### Multiple nodes, multiple GPUs
 
 Let's start training on two nodes with 2 gpus each. We assuming that master node can be connected as `master`, e.g. `ping master`.
 
-1) Execute on master node
+1. Execute on master node
+
 ```bash
 python -u -m torch.distributed.launch \
     --nnodes=2 \
@@ -90,7 +98,8 @@ python -u -m torch.distributed.launch \
     main.py run --backend="nccl"
 ```
 
-2) Execute on worker node
+2. Execute on worker node
+
 ```bash
 python -u -m torch.distributed.launch \
     --nnodes=2 \
@@ -100,17 +109,18 @@ python -u -m torch.distributed.launch \
     main.py run --backend="nccl"
 ```
 
-
 ### Check resume training
 
 #### Single GPU
 
 Initial training with a stop on 1000 iteration (~11 epochs)
+
 ```bash
 python main.py run --stop_iteration=1000
 ```
 
 Resume from the latest checkpoint
+
 ```bash
 python main.py run --resume-from=/tmp/output-cifar10/resnet18_backend-None-1_stop-on-1000/training_checkpoint_1000.pt
 ```
@@ -120,12 +130,14 @@ python main.py run --resume-from=/tmp/output-cifar10/resnet18_backend-None-1_sto
 #### Single node, multiple GPUs
 
 Initial training on a single node with 2 gpus with a stop on 1000 iteration (~11 epochs):
+
 ```bash
 # using torch.distributed.launch
 python -u -m torch.distributed.launch --nproc_per_node=2 --use_env main.py run --backend="nccl" --stop_iteration=1000
 ```
 
 Resume from the latest checkpoint
+
 ```bash
 python -u -m torch.distributed.launch --nproc_per_node=2 --use_env main.py run --backend="nccl" \
     --resume-from=/tmp/output-cifar10/resnet18_backend-nccl-2_stop-on-1000/training_checkpoint_1000.pt
@@ -133,12 +145,12 @@ python -u -m torch.distributed.launch --nproc_per_node=2 --use_env main.py run -
 
 Similar commands can be adapted for other cases.
 
-## Trains fileserver
+## ClearML fileserver
 
-If `Trains` server is used (i.e. `--with_trains` argument), the configuration to upload artifact must be done by 
-modifying the `Trains` configuration file `~/trains.config` generated by `trains-init`. According to the
-[documentation](https://allegro.ai/docs/examples/reporting/artifacts/), the `output_uri` argument can be 
-configured in `sdk.development.default_output_uri` to fileserver uri. If server is self-hosted, `Trains` fileserver uri is
+If `ClearML` server is used (i.e. `--with_clearml` argument), the configuration to upload artifact must be done by
+modifying the `ClearML` configuration file `~/clearml.conf` generated by `clearml-init`. According to the
+[documentation](https://allegro.ai/clearml/docs/docs/examples/reporting/artifacts.html), the `output_uri` argument can be
+configured in `sdk.development.default_output_uri` to fileserver uri. If server is self-hosted, `ClearML` fileserver uri is
 `http://localhost:8081`.
 
-For more details, see https://allegro.ai/docs/examples/reporting/artifacts/
+For more details, see https://allegro.ai/clearml/docs/docs/examples/reporting/artifacts.html
