+++ examples/references/classification/imagenet/experiments/clearml/requirements.txt
@@ -1,0 +1,1 @@
+clearml
