--- examples/references/classification/imagenet/experiments/trains/requirements.txt
@@ -1,1 +1,0 @@
-trains >= 0.15.0
