+++ setup.cfg
@@ -7,7 +7,7 @@ ignore = E402, E721
 max_line_length = 120
 
 [isort]
-known_third_party=dill,matplotlib,numpy,pkg_resources,pytest,requests,setuptools,skimage,sklearn,torch,torchvision,trains
+known_third_party=clearml,dill,matplotlib,numpy,pkg_resources,pytest,requests,setuptools,skimage,sklearn,torch,torchvision
 multi_line_output=3
 include_trailing_comma=True
 force_grid_wrap=0
